refactor: update package structure for SuggestEasyMeal components
