+++ src/main/kotlin/di/UseCaseModule.kt
@@ -5,10 +5,7 @@ import logic.usecase.filter.GymMealHelperUseCase
 import org.example.logic.usecase.filter.SeaFoodMealUseCase
 import org.example.logic.usecase.game.GuessMealGameUseCase
 import org.example.logic.usecase.game.IngredientGameUseCase
-import org.example.logic.usecase.suggest.EggFreeSweetsUseCase
-import org.example.logic.usecase.suggest.IraqiMealsIdentifierUseCase
-import org.example.logic.usecase.suggest.MealsForLargeGroupUseCase
-import org.example.logic.usecase.suggest.SuggestHighCalorieMealUseCase
+import org.example.logic.usecase.suggest.*
 import org.koin.dsl.module
 
 val useCaseModule = module {
